Created menu control package and move more classes into the view package.
