+++ battleship/src/citbyui260/section03/battleship/view/GameMenu.java
@@ -3,8 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package battleship;
+package citbyui260.section03.battleship.view;
 
+import battleship.BattleshipError;
+import battleship.Game;
+import citbyui260.section03.battleship.control.GameMenuControl;
+import citbyui260.section03.battleship.view.MenuSuper;
 import java.io.Serializable;
 import java.util.Scanner;
