+++ battleship/src/citbyui260/section03/battleship/view/Help.java
@@ -6,8 +6,7 @@
 package citbyui260.section03.battleship.view;
 
 import battleship.BattleshipError;
-import battleship.HelpMenuControl;
-import battleship.MenuSuper;
+import citbyui260.section03.battleship.control.HelpMenuControl;
 import java.util.Scanner;
