+++ battleship/src/citbyui260/section03/battleship/view/Menu.java
@@ -3,8 +3,11 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package battleship;
+package citbyui260.section03.battleship.view;
 
+import battleship.BattleshipError;
+import citbyui260.section03.battleship.control.MenuControl;
+import citbyui260.section03.battleship.view.MenuSuper;
 import java.util.Scanner;
 /**
  *
