+++ battleship/src/citbyui260/section03/battleship/view/MenuSuper.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package battleship;
+package citbyui260.section03.battleship.view;
 import citbyui260.section03.battleship.interfaces.*;
 
 /**
